// src/modules/contact/types.ts
export interface Props {
  translations: {
    fullname: string
    fullnamePlaceholder: string
    phoneNumber: string
    email: string
    message: string
    messagePlaceholder: string
    send: string
    cleanForm: string
    errorEmptyField: string
    errorInvalidEmail: string
    submitSuccess: string
    submitError: string
  }
}